Moving Slate spark classes to new package.
